core/pipe: remove single-process implementation of Pipe
benchmarks show there is only a marginal performance difference in this
implementation compared to the multiprocess POSIX pipe() on heavy
compute-bound pipelines. also, the POSIX pipe() implementation supports
partial reads, which can be necessary in blocks processing more than one
input from sources with disparate throughputs, so this change brings
support for those pipelines to singleprocess execution.

in the future, a more efficient singleprocess implementation may return,
if the multiple inputs issue is addressed for it.