blocks/sources/signal: fix numerical stability of exp, cos, sin generators
previously, these sinuosoids were generated by rotating a unit phase
vector around the unit circle by multiplying with a unit rotator vector.
the problem was that the rotator does not have a magnitude of precisely
1, because of limited float32 precision, so over time the phase vector
magnitude would either shrink to zero or grow to infinity.

this commit changes the generation to manually increment a scalar phase
modulo 2pi, and use math.cos() and/or math.sin() to compute each point
of the signal.